Cambios en la parte del modelo y se agrego el archivo ignore para que no se agreguen archivos que no tienen que estar

+++ TusAutos/Models/Auto.cs
@@ -42,8 +42,7 @@ namespace TusAutos.Models
 
         [DisplayName("Consesionaria")]
         [ForeignKey("Consesionaria")]
-        
-        public long ConsesionariaId { get; set; }
+        public int ConsesionariaId { get; set; }
         public virtual Consesionaria Consesionaria { get; set; }
 
         public virtual ICollection<Imagen> Imagenes { get; set; }
